+++ pkg/types/cluster.go
@@ -16,15 +16,15 @@ type Cluster struct {
 	NodesCount     int    `json:"nodeCount,omitempty"`
 	Version        string `json:"version,omitempty"`
 
-	Cpu             float64 `json:"cpu,omitempty"`
-	CpuUsed         float64 `json:"-"`
-	CpuUsedRatio    string  `json:"cpuUsedRatio,omitempty"`
-	Memory          float64 `json:"memory,omitempty"`
-	MemoryUsed      float64 `json:"-"`
-	MemoryUsedRatio string  `json:"memoryUsedRatio,omitempty"`
-	Pod             float64 `json:"pod"`
-	PodUsed         float64 `json:"-"`
-	PodUsedRatio    string  `json:"podUsedRatio,omitempty"`
+	Cpu             int64  `json:"cpu,omitempty"`
+	CpuUsed         int64  `json:"-"`
+	CpuUsedRatio    string `json:"cpuUsedRatio,omitempty"`
+	Memory          int64  `json:"memory,omitempty"`
+	MemoryUsed      int64  `json:"-"`
+	MemoryUsedRatio string `json:"memoryUsedRatio,omitempty"`
+	Pod             int64  `json:"pod"`
+	PodUsed         int64  `json:"-"`
+	PodUsedRatio    string `json:"podUsedRatio,omitempty"`
 }
 
 var ClusterType = types.GetResourceType(Cluster{})
